Fix string replace bug & drop nan rows
The NSPN_rois_behavmerge code has for a while written out the name of
the input file as a suffix to all the combined data, but this commit
fixes a little bug that replaced all instances of c s or v rather than
just the ".csv" string at the end.

The comment for this step has also been updated - it was a left over
from when we added _behavmerge.csv.

Finally, this commit drops rows in the behavioural data file that are
missing an nspn id and therefore fixes an error that was thrown by the
merge step (it couldn't merge NaNs).